Created constants file
