+++ task-manager/src/utils/constants.ts
@@ -1,0 +1,13 @@
+export enum OPERATIONS {
+  VIEW = "View",
+  ADD = "Add",
+  EDIT = "Edit",
+  STORE = "Store",
+}
+
+export const STORAGE_KEY = "TASKS";
+
+export enum TASK_STATUS {
+  COMPLETED = "Completed",
+  INCOMPLETE = "Incomplete",
+}
